+++ source/views/menus/menu-bonapp.js
@@ -1,6 +1,5 @@
 // @flow
 import * as React from 'react'
-import delay from 'delay'
 import {timezone} from '@frogpond/constants'
 import {NoticeView, LoadingView} from '@frogpond/notice'
 import type {TopLevelViewPropsType} from '../types'
@@ -106,11 +105,11 @@ export class BonAppHostedMenu extends React.PureComponent<Props, State> {
 
 		try {
 			let cafeMenuPromise: Promise<MenuInfoType> = fetch(menuUrl, {
-				forReload: reload ? 500 : 0,
+				delay: reload ? 500 : 0,
 			}).json()
 
 			let cafeInfoPromise: Promise<CafeInfoType> = fetch(cafeUrl, {
-				forReload: reload ? 500 : 0,
+				delay: reload ? 500 : 0,
 			}).json()
 
 			let [cafeMenu, cafeInfo] = await Promise.all([
@@ -133,17 +132,8 @@ export class BonAppHostedMenu extends React.PureComponent<Props, State> {
 	}
 
 	refresh = async (): any => {
-		let start = Date.now()
 		this.setState(() => ({refreshing: true}))
-
 		await this.fetchData(this.props, true)
-
-		// wait 0.5 seconds – if we let it go at normal speed, it feels broken.
-		let elapsed = Date.now() - start
-		if (elapsed < 500) {
-			await delay(500 - elapsed)
-		}
-
 		this.setState(() => ({refreshing: false}))
 	}